Introduce common/config_utils.py
