+++ admin/server/config.py
@@ -21,7 +21,7 @@ from enum import Enum
 
 from pydantic import BaseModel
 from typing import Any
-from api.utils.configs import read_config
+from common.config_utils import read_config
 from urllib.parse import urlparse
