+++ deepdoc/parser/tcadp_parser.py
@@ -36,7 +36,7 @@ from tencentcloud.common.profile.http_profile import HttpProfile
 from tencentcloud.common.exception.tencent_cloud_sdk_exception import TencentCloudSDKException
 from tencentcloud.lkeap.v20240522 import lkeap_client, models
 
-from api.utils.configs import get_base_config
+from common.config_utils import get_base_config
 from deepdoc.parser.pdf_parser import RAGFlowPdfParser
