+++ rag/svr/task_executor.py
@@ -30,7 +30,7 @@ from api.utils.api_utils import timeout
 from common.base64_image import image2id
 from api.utils.log_utils import init_root_logger
 from common.file_utils import get_project_base_directory
-from api.utils.configs import show_configs
+from common.config_utils import show_configs
 from graphrag.general.index import run_graphrag_for_kb
 from graphrag.utils import get_llm_cache, set_llm_cache, get_tags_from_cache, set_tags_to_cache
 from rag.flow.pipeline import Pipeline
